# neverland/internal/docs/static/openapi.yaml
openapi: 3.1.0
info:
  title: Neverland
  description: Tinkerbell management gateway — REST API for bare-metal provisioning
  version: 0.1.0
servers:
  - url: http://10.11.30.91:8091
    description: LAN (k3s node)
paths:
...
  /api/v1/artifacts/fetch:
    get:
      summary: List all fetch jobs
      operationId: listFetchJobs
      tags: [Artifacts]
      responses:
        "200":
          description: List of fetch jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/FetchJob"
    post:
      summary: Fetch and convert an image from a URL into a Tinkerbell-ready raw.gz artifact
      description: |
        Downloads the image at `url`, detects its format, converts qcow2 → raw if needed,
        then gzip-compresses it to `<name>.raw.gz` in the artifacts directory.
        Returns immediately with a job ID — poll `GET /api/v1/artifacts/fetch/{id}` for status.

        **Compatible source formats:** `.raw`, `.raw.gz` (passthrough), `.qcow2` (converted), `.img`
        **Output:** always `.raw.gz` — plug the `outputURL` directly into a workflow's `IMG_URL`
      operationId: startFetchJob
      tags: [Artifacts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchRequest"
      responses:
        "202":
          description: Job accepted — poll for status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchJob"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/artifacts/fetch/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get fetch job status
      operationId: getFetchJob
      tags: [Artifacts]
      responses:
        "200":
          description: Fetch job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchJob"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    FetchRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: HTTP URL of the source image (qcow2, raw, img, or raw.gz)
          example: "https://cloud-images.ubuntu.com/noble/current/noble-server-cloudimg-amd64.img"
        name:
          type: string
          description: Output filename (without extension — .raw.gz will be appended). Defaults to filename from URL.
          example: ubuntu-24.04-server
    FetchJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        name:
          type: string
          example: ubuntu-24.04-server.raw.gz
        status:
          type: string
          enum: [pending, downloading, converting, compressing, done, error]
        message:
          type: string
        outputURL:
          type: string
          description: Nginx URL to use as IMG_URL in a workflow — only set when status is done
          example: "http://10.11.30.91:8080/ubuntu-24.04-server.raw.gz"
        startedAt:
          type: string
          format: date-time
        doneAt:
          type: string
          format: date-time
...